requirements.txt
diffusers
transformers
accelerate
torch
requests
python-dotenv
Pillow
google-genai
